fix FFA biped color showing white in game_engine_player_get_change_color
The forced-color-override check at 0x2efe20 was read as int (32-bit) but
the original uses a 16-bit CMP (CMP CX, -1 at 0xa99e9). In FFA with no
override active the field is 0xffff (int16 = -1), but adjacent bytes are
0x0000, so the 32-bit read yielded 0x0000ffff != -1, falsely triggering
the override path with color_index = -1. FUN_001c0ee0(-1) returns white,
causing all FFA bipeds to appear white instead of their profile color.

Fix: compare as int16_t to match the original CMP CX, -1.

+++ src/halo/game/game_engine.c
@@ -5469,7 +5469,7 @@ float *game_engine_player_get_change_color(float *param_1, int param_2)
 
     color_index = *(int16_t *)(player + 0x60);
 
-    if (*(int *)0x2efe20 != -1)
+    if (*(int16_t *)0x2efe20 != -1)
 
       color_index = *(int16_t *)0x2efe20;
